requirements.txt
Django>=4.0,<5.0
djangorestframework
requests
Pillow
transformers
torch
torchvision
numpy
matplotlib
opencv-python
python-dotenv
gunicorn
whitenoise==6.5.0